+++ src/leveldb/db/repair.cc
@@ -75,7 +75,7 @@ class Repairer {
       status = WriteDescriptor();
     }
     if (status.ok()) {
-      unsigned long long bytes = 0;
+      uint64_t bytes = 0;
       for (size_t i = 0; i < tables_.size(); i++) {
         bytes += tables_[i].meta.file_size;
       }
@@ -152,7 +152,7 @@ class Repairer {
       Status status = ConvertLogToTable(logs_[i]);
       if (!status.ok()) {
         Log(options_.info_log, "Log #%llu: ignoring conversion error: %s",
-            (unsigned long long) logs_[i],
+            (uint64_t) logs_[i],
             status.ToString().c_str());
       }
       ArchiveFile(logname);
@@ -167,7 +167,7 @@ class Repairer {
       virtual void Corruption(size_t bytes, const Status& s) {
         // We print error messages for corruption, but continue repairing.
         Log(info_log, "Log #%llu: dropping %d bytes; %s",
-            (unsigned long long) lognum,
+            (uint64_t) lognum,
             static_cast<int>(bytes),
             s.ToString().c_str());
       }
@@ -212,7 +212,7 @@ class Repairer {
         counter += WriteBatchInternal::Count(&batch);
       } else {
         Log(options_.info_log, "Log #%llu: ignoring %s",
-            (unsigned long long) log,
+            (uint64_t) log,
             status.ToString().c_str());
         status = Status::OK();  // Keep going with rest of file
       }
@@ -234,9 +234,9 @@ class Repairer {
       }
     }
     Log(options_.info_log, "Log #%llu: %d ops saved to Table #%llu %s",
-        (unsigned long long) log,
+        (uint64_t) log,
         counter,
-        (unsigned long long) meta.number,
+        (uint64_t) meta.number,
         status.ToString().c_str());
     return status;
   }
@@ -272,7 +272,7 @@ class Repairer {
       ArchiveFile(TableFileName(dbname_, number));
       ArchiveFile(SSTTableFileName(dbname_, number));
       Log(options_.info_log, "Table #%llu: dropped: %s",
-          (unsigned long long) t.meta.number,
+          (uint64_t) t.meta.number,
           status.ToString().c_str());
       return;
     }
@@ -287,7 +287,7 @@ class Repairer {
       Slice key = iter->key();
       if (!ParseInternalKey(key, &parsed)) {
         Log(options_.info_log, "Table #%llu: unparsable key %s",
-            (unsigned long long) t.meta.number,
+            (uint64_t) t.meta.number,
             EscapeString(key).c_str());
         continue;
       }
@@ -307,7 +307,7 @@ class Repairer {
     }
     delete iter;
     Log(options_.info_log, "Table #%llu: %d entries %s",
-        (unsigned long long) t.meta.number,
+        (uint64_t) t.meta.number,
         counter,
         status.ToString().c_str());
 
@@ -363,7 +363,7 @@ class Repairer {
       s = env_->RenameFile(copy, orig);
       if (s.ok()) {
         Log(options_.info_log, "Table #%llu: %d entries repaired",
-            (unsigned long long) t.meta.number, counter);
+            (uint64_t) t.meta.number, counter);
         tables_.push_back(t);
       }
     }
